+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 (YYYY-MM-DD):
 
 * Using module koanf to read and parse the yaml configuration file
+* Instances and volumes can be filtered using more than one tag with a new syntax
 
 ## 0.1.0 (2024-01-18):